rework data-buffer-file
