+++ src/src.pro
@@ -11,7 +11,6 @@ INCLUDEPATH += $$PWD \
 SOURCES += \
     threading/thread.cpp \
     statemachine.cpp \
-    data_buffer_methods.cpp \
     common_items/data_items.cpp \
     common_items/table_item.cpp \
     unit_test.cpp \
@@ -20,7 +19,6 @@ SOURCES += \
 
 HEADERS += \
     ../include/libKitsunemimiCommon/data_buffer.h \
-    ../include/libKitsunemimiCommon/data_buffer_methods.h \
     ../include/libKitsunemimiCommon/common_methods/string_methods.h \
     ../include/libKitsunemimiCommon/common_methods/vector_methods.h \
     ../include/libKitsunemimiCommon/common_items/data_items.h \
