+++ tests/libKitsunemimiCommon/data_buffer_test.cpp
@@ -29,6 +29,10 @@ DataBuffer_Test::DataBuffer_Test()
     addData_test();
     getBlock_test();
     reset_test();
+
+    addDataToBuffer_test();
+    allocateBlocks_test();
+    resetBuffer_test();
 }
 
 /**
@@ -152,6 +156,97 @@ DataBuffer_Test::reset_test()
     TEST_EQUAL(testBuffer.numberOfBlocks, 1);
     TEST_EQUAL(testBuffer.bufferPosition, 0);
     TEST_EQUAL(testBuffer.totalBufferSize, 1*testBuffer.blockSize);
+
+    // check content of the buffer
+    uint8_t* dataByte = static_cast<uint8_t*>(testBuffer.data);
+    TEST_EQUAL(static_cast<int>(dataByte[1]), 0);
+}
+
+/**
+ * addDataToBuffer_test
+ */
+void
+DataBuffer_Test::addDataToBuffer_test()
+{
+    // init
+    DataBuffer testBuffer(10);
+    TestStruct testStruct;
+    testStruct.b = 42;
+
+    // check metadata of the buffer
+    TEST_EQUAL(testBuffer.bufferPosition, 0);
+
+    // add data to buffer
+    void* testStructPtr = static_cast<void*>(&testStruct);
+    TEST_EQUAL(addDataToBuffer(&testBuffer, testStructPtr, sizeof(TestStruct)), true);
+
+    // check metadata of the buffer
+    TEST_EQUAL(testBuffer.numberOfBlocks, 10);
+    TEST_EQUAL(testBuffer.bufferPosition, 10);
+    TEST_EQUAL(testBuffer.totalBufferSize, 10*testBuffer.blockSize);
+
+    // check content of the buffer
+    uint8_t* dataByte = static_cast<uint8_t*>(testBuffer.data);
+    TEST_EQUAL(static_cast<int>(dataByte[1]), 42);
+}
+
+/**
+ * allocateBlocks_test
+ */
+void
+DataBuffer_Test::allocateBlocks_test()
+{
+    // init
+    DataBuffer testBuffer(10);
+    TestStruct testStruct;
+    testStruct.b = 42;
+
+    // write data to buffer
+    TEST_EQUAL(testBuffer.addData(&testStruct), true);
+
+    // check metadata of the buffer
+    TEST_EQUAL(testBuffer.numberOfBlocks, 10);
+    TEST_EQUAL(testBuffer.bufferPosition, 10);
+    TEST_EQUAL(testBuffer.totalBufferSize, 10*testBuffer.blockSize);
+
+    // check content of the buffer
+    uint8_t* dataByte = static_cast<uint8_t*>(testBuffer.data);
+    TEST_EQUAL(static_cast<int>(dataByte[1]), 42);
+
+    // resize
+    TEST_EQUAL(allocateBlocks(&testBuffer, 1), true);
+
+    // check metadata of the buffer
+    TEST_EQUAL(testBuffer.numberOfBlocks, 11);
+    TEST_EQUAL(testBuffer.bufferPosition, 10);
+    TEST_EQUAL(testBuffer.totalBufferSize, 11*testBuffer.blockSize);
+
+    // check content of the buffer
+    dataByte = static_cast<uint8_t*>(testBuffer.data);
+    TEST_EQUAL(static_cast<int>(dataByte[1]), 42);
+}
+
+/**
+ * resetBuffer_test
+ */
+void
+DataBuffer_Test::resetBuffer_test()
+{
+    // init
+    DataBuffer testBuffer(10);
+    TestStruct testStruct;
+    testStruct.b = 42;
+
+    // write data to buffer
+    TEST_EQUAL(testBuffer.addData(&testStruct), true);
+
+    // reset buffer
+    TEST_EQUAL(resetBuffer(&testBuffer, 2), true);
+
+    // check metadata of the buffer
+    TEST_EQUAL(testBuffer.numberOfBlocks, 2);
+    TEST_EQUAL(testBuffer.bufferPosition, 0);
+    TEST_EQUAL(testBuffer.totalBufferSize, 2*testBuffer.blockSize);
 
     // check content of the buffer
     uint8_t* dataByte = static_cast<uint8_t*>(testBuffer.data);
