+++ tests/tests.pro
@@ -13,7 +13,6 @@ HEADERS += \
     libKitsunemimiCommon/statemachine_test.h \
     libKitsunemimiCommon/common_methods/string_methods_test.h \
     libKitsunemimiCommon/common_methods/vector_methods_test.h \
-    libKitsunemimiCommon/data_buffer_methods_test.h \
     libKitsunemimiCommon/data_buffer_test.h \
     libKitsunemimiCommon/common_items/data_items_DataArray_test.h \
     libKitsunemimiCommon/common_items/data_items_DataValue_test.h \
@@ -27,7 +26,6 @@ SOURCES += \
     libKitsunemimiCommon/data_buffer_test.cpp \
     libKitsunemimiCommon/common_methods/string_methods_test.cpp \
     libKitsunemimiCommon/common_methods/vector_methods_test.cpp \
-    libKitsunemimiCommon/data_buffer_methods_test.cpp \
     libKitsunemimiCommon/common_items/data_items_DataArray_test.cpp \
     libKitsunemimiCommon/common_items/data_items_DataValue_test.cpp \
     libKitsunemimiCommon/common_items/table_item_test.cpp \
